Adding comments and minor simplifications to get rid of some more #if blocks.
