+++ SignalR.Client.Net20/SignalR.Client.Net20.csproj
@@ -132,9 +132,14 @@
       <Link>Infrastructure\UriQueryUtility.cs</Link>
     </Compile>
     <Compile Include="Infrastructure\CancellationTokenSource.cs" />
+    <Compile Include="Infrastructure\CustomResultArgs.cs" />
     <Compile Include="Infrastructure\IObservable.cs" />
+    <Compile Include="Infrastructure\ReadStreamState.cs" />
+    <Compile Include="Infrastructure\ResultWrapper.cs" />
     <Compile Include="Infrastructure\StreamExtensions.cs" />
     <Compile Include="Infrastructure\Task.cs" />
+    <Compile Include="Infrastructure\TaskAsyncHelper.cs" />
+    <Compile Include="Infrastructure\WriteStreamState.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\VersionInfo.cs" />
   </ItemGroup>
